Make a start with defining matchable functions
Specifically, I am starting with injectivity via `primInjArrow`.
We need to turn the body into the appropriate implication and add
the instances for lambdas being functions to avoid the type errors.

Next, we can define generativity (which should be very similar) and
finally finish with matchability as the conjunction of both.